+++ docs/changelog.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.2.2]
+### Fixed
+ - Back compatibility with 0.2.0 version of the client
+
 ## [0.2.1]
 ### Added
  - Add opportunity to create elements by scs